firmware: app: tasks: Include system.h in read sensors task

+++ firmware/app/tasks/read_sensors.c
@@ -36,6 +36,7 @@
  */
 
 #include <system/sys_log/sys_log.h>
+#include <system/system.h>
 
 #include <devices/current_sensor/current_sensor.h>
 #include <devices/voltage_sensor/voltage_sensor.h>
@@ -483,7 +484,7 @@ void vTaskReadSensors(void)
         }
 
         /* Update timestamp */
-        uint32_t time_counter = (uint32_t)system_get_time();
+        uint32_t time_counter = system_get_time();
         eps_buffer_write(EPS2_PARAM_ID_TIMESTAMP, &time_counter);
 
         vTaskDelayUntil(&last_cycle, pdMS_TO_TICKS(TASK_READ_SENSORS_PERIOD_MS));
